refactor: more css changes and memo instead of effect
