+++ apis/nucleus/src/components/listbox/components/ListBoxRowColumn/components/ItemGrid.jsx
@@ -4,9 +4,12 @@ import getItemSizes from '../../grid-list-components/item-sizes';
 import classes from '../helpers/classes';
 
 const ItemGrid = styled(Grid, {
-  shouldForwardProp: (prop) => !['dataLayout', 'layoutOrder', 'itemPadding'].includes(prop),
-})(({ dataLayout, layoutOrder, itemPadding }) => ({
-  [`&.${classes.fieldRoot}`]: getItemSizes({ dataLayout, layoutOrder, itemPadding }),
+  shouldForwardProp: (prop) => !['dataLayout', 'layoutOrder', 'itemPadding', 'checkboxes'].includes(prop),
+})(({ dataLayout, layoutOrder, itemPadding, checkboxes }) => ({
+  [`&.${classes.fieldRoot}`]: {
+    ...getItemSizes({ dataLayout, layoutOrder, itemPadding }),
+    paddingRight: checkboxes ? '8px' : undefined,
+  },
 }));
 
 export default ItemGrid;
